add auth coordinator
